feat: Adding lightbox with amenities details
