+++ src/layouts/index/amenities/Amenities.module.sass
@@ -5,8 +5,6 @@
     flex-direction: column
     justify-content: center
     align-items: center
-    width: $default-container-width
-    max-width: 1200px
     row-gap: 40px
 
     .gallery
@@ -15,12 +13,5 @@
         grid-template-columns: 1fr 1fr
         grid-auto-rows: auto
         gap: 32px
-
-        .item
-            position: relative
-            height: 250px
-
-            .thumb
-                position: relative
-                height: 100%
-                width: 100%
+        width: $default-container-width
+        max-width: 1200px
