--- src/layouts/index/amenities/detail/Detail.tsx
@@ -1,5 +1,0 @@
-export type Props = {
-    title: string
-    image: string
-    description: string
-}
