+++ src/layouts/index/amenities/detailwiththumb/DetailWithThumb.module.sass
@@ -1,0 +1,36 @@
+@import @styles/constants
+
+.item
+    height: 250px
+
+    .thumb
+        position: relative
+        height: 100%
+        width: 100%
+
+    .lightbox
+        position: absolute
+        top: 0
+        left: 0
+        width: 100%
+        height: 100%
+        background-color: $color-background
+
+        .back
+            position: absolute
+            bottom: 0
+            left: 48px
+            padding: 0
+            background: 0
+            border: 0
+            text-transform: uppercase
+            font-family: $font-aviano-flare
+            font-weight: bold
+            color: $color-primary
+
+            &::before
+                content: "< "
+
+            &:hover
+                text-decoration: underline
+                cursor: pointer
